requirements.txt
certifi==2019.6.16
decorator==4.4.0
imageio==2.5.0
imutils==0.5.3
joblib==0.13.2
networkx==2.3
numpy==1.17.2
opencv-python==4.1.1.26
Pillow==6.1.0
PyWavelets==1.0.3
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
wincertstore==0.2
tqdm==4.36.1